refactor: move CannotModifyPublishedRecipe error class to separate file
