+++ src/main/scala/api/recipes/ingredients/CannotModifyPublishedRecipe.scala
@@ -1,0 +1,16 @@
+package api.recipes.ingredients
+
+import api.variantJson
+import domain.RecipeId
+
+import io.circe.generic.auto.*
+import sttp.model.StatusCode.Forbidden
+import sttp.tapir.generic.auto.*
+
+final case class CannotModifyPublishedRecipe(
+  recipeId: RecipeId,
+  messae: String = "Cannot modify published recipe",
+)
+
+object CannotModifyPublishedRecipe:
+  val variant = Forbidden.variantJson[CannotModifyPublishedRecipe]
